requirements.txt
numpy>=1.24.2
Pillow>=9.4.0
